+++ src/main/java/nextstep/session/domain/student/SessionStudents.java
@@ -1,0 +1,76 @@
+package nextstep.session.domain.student;
+
+import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+public class SessionStudents {
+    private Set<SessionStudent> students;
+    private int maximumCapacity;
+
+    public SessionStudents(Set<SessionStudent> students, int maximumCapacity) {
+        this.students = students;
+        this.maximumCapacity = maximumCapacity;
+    }
+
+    public SessionStudents(List<SessionStudent> students, int maximumCapacity) {
+        this.students = students.stream().collect(Collectors.toSet());
+        this.maximumCapacity = maximumCapacity;
+    }
+
+    public SessionStudents(int maximumCapacity) {
+        this(new HashSet<>(), maximumCapacity);
+    }
+
+    private boolean canRecruitStudent(int maximumCapacity) {
+        return maximumCapacity <= students.size();
+    }
+
+    public boolean contains(SessionStudent student) {
+        return students.contains(student);
+    }
+
+    public void enrollStudent(SessionStudent student) {
+        System.out.println("students = " + students);
+        if (contains(student)) {
+            throw new IllegalArgumentException("이미 등록된 학생입니다");
+        }
+
+        if (canRecruitStudent(maximumCapacity)) {
+            throw new IllegalArgumentException("정원수를 초과했습니다");
+        }
+
+        students.add(student);
+    }
+
+    public int getCurrentStudentCount() {
+        return students.size();
+    }
+
+    public int getMaximumCapacity() {
+        return maximumCapacity;
+    }
+
+    @Override
+    public String toString() {
+        return "SessionStudents{" +
+                "students=" + students +
+                ", maximumCapacity=" + maximumCapacity +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        SessionStudents that = (SessionStudents) o;
+        return maximumCapacity == that.maximumCapacity && Objects.equals(students, that.students);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(students, maximumCapacity);
+    }
+}
