--- src/main/java/nextstep/sessions/domain/SessionPeriod.java
@@ -1,51 +1,0 @@
-package nextstep.sessions.domain;
-
-import java.time.LocalDateTime;
-import java.util.Objects;
-
-public class SessionPeriod {
-
-    private LocalDateTime startDate;
-    private LocalDateTime endDate;
-
-    public SessionPeriod(LocalDateTime startDate, LocalDateTime endDate) {
-        validateDates(startDate, endDate);
-        this.startDate = startDate;
-        this.endDate = endDate;
-    }
-
-    public LocalDateTime getStartDate() {
-        return startDate;
-    }
-
-    public LocalDateTime getEndDate() {
-        return endDate;
-    }
-
-    private void validateDates(LocalDateTime startDate, LocalDateTime endDate) {
-        if (startDate.isAfter(endDate)) {
-            throw new IllegalArgumentException("시작 날짜가 종료 날짜보다 더 늦습니다");
-        }
-    }
-
-    @Override
-    public String toString() {
-        return "SessionPeriod{" +
-                "startDate=" + startDate +
-                ", endDate=" + endDate +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        SessionPeriod that = (SessionPeriod) o;
-        return Objects.equals(startDate, that.startDate) && Objects.equals(endDate, that.endDate);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(startDate, endDate);
-    }
-}
