+++ src/main/resources/schema.sql
@@ -61,7 +61,8 @@ create table session
     start_date       timestamp,
     end_date         timestamp,
     payment_type     varchar(255),
-    session_status   varchar(255),
+    progress_status   varchar(255),
+    recruitment_status   varchar(255),
     session_capacity bigint,
     created_at       timestamp not null,
     updated_at       timestamp,
@@ -70,16 +71,30 @@ create table session
 
 create table session_student
 (
-    session_id bigint not null,
-    ns_user_id bigint not null,
-    primary key (session_id, ns_user_id)
+    id         bigint generated by default as identity,
+    session_id bigint         not null,
+    ns_user_id bigint         not null,
+    status     character(255) not null,
+    created_at timestamp      not null,
+    updated_at timestamp,
+    primary key (id)
+);
+
+create table session_teacher
+(
+    id         bigint generated by default as identity,
+    session_id bigint    not null,
+    ns_user_id bigint    not null,
+    created_at timestamp not null,
+    updated_at timestamp,
+    primary key (id)
 );
 
 create table cover_image
 (
     id         bigint generated by default as identity,
     session_id bigint    not null,
-    image_name  varchar(255),
+    image_name varchar(255),
     image_type varchar(255),
     url        varchar(255),
     created_at timestamp not null,
