--- src/test/java/nextstep/sessions/domain/SessionCoverImageTest.java
@@ -1,14 +1,0 @@
-package nextstep.sessions.domain;
-
-public class SessionCoverImageTest {
-    public static final SessionCoverImage image1 = new SessionCoverImage(
-            1L,
-            "image_name1",
-            "https://blog.kakaocdn.net/dn/da8XaC/btsh40VDDI5/ZQmYKwA6Ky7EKkL9GK3A2K/img.jpg",
-            10000000, 10000000);
-    public static final SessionCoverImage image3 = new SessionCoverImage(
-            3L,
-            "image_name3",
-            "https://blog.kakaocdn.net/dn/da8XaC/btsh40VDDI5/ZQmYKwA6Ky7EKkL9GK3A2K/img.jpg",
-            10000000, 10000000);
-}
